Implemented workflows defined with group 4
