+++ logic/workflow.py
@@ -1,4 +1,6 @@
+import json
 from message import Message, Method, State, fromJSON
+from enum import Enum
 
 
 class Workflow:
@@ -8,7 +10,7 @@ class Workflow:
     https://github.com/ubilab-escape/operator/blob/master/doc/design/general_%CE%BCC_workflow.svg
     """
 
-    def __init__(self, name, topic):
+    def __init__(self, name, topic, settings=None):
         """
         Initializes a new instance of this class.
 
@@ -16,11 +18,16 @@ class Workflow:
         ----------
         name : str
             Display name of the workflow.
+
         topic : str
             Name of the MQTT topic.
+
+        settings: keywords
+            An dictionary of global settings.
         """
         self.topic = topic
         self.name = name
+        self.settings = settings
         self._on_workflow_failed = None
         self._on_workflow_solved = None
 
@@ -33,7 +40,8 @@ class Workflow:
         client : Client
             MQTT client
         """
-        message = Message(Method.TRIGGER, State.ON)
+        data = self.getData(self.settings)
+        message = Message(Method.TRIGGER, State.ON, data)
         client.publish(self.topic, message.toJSON())
         print("[%s] Started..." % (self.name))
         client.subscribe(self.topic)
@@ -50,6 +58,16 @@ class Workflow:
         """
         client.unsubscribe(self.topic)
         print("[%s] Unsubscribed from topic '%s'..." % (self.name, self.topic))
+
+    def getData(self, settings):
+        data = None
+
+        if settings:
+            if len(settings) == 1:
+                data = next(iter(settings.values()))
+            else:
+                data = json.dumps(settings)
+        return data
 
     def on_message(self, msg):
         """
@@ -149,3 +167,116 @@ class Workflow:
 
     def on_received_trigger_off(self, data):
         print("  ==> Nothing to do")
+
+
+class ParallelWorkflow(Workflow):
+    """
+    This class implements a wrapper to run multiple workflows in parallel.
+    The parallel workflow is a composition organising the flow of one or
+    more arbitary workflows.
+    """
+
+    def __init__(self, name, workflows):
+        """
+        Initializes a new instance of this class.
+
+        Parameters
+        ----------
+        name : str
+            Display name of the workflow.
+        topic : str
+            Name of the MQTT topic.
+        workflows : Workflow[]
+            Collection of workflows should be executed in parallel.
+        """
+        super().__init__(name, None)
+        self.workflows = workflows
+        self.workflow_solved = {}
+        for workflow in self.workflows:
+            self.workflow_solved[workflow.name] = False
+            workflow.register_on_failed(self.__on_workflow_failed)
+            workflow.register_on_solved(self.__on_workflow_solved)
+
+    def execute(self, client):
+        """
+        Executes this workflow.
+
+        Parameters
+        ----------
+        client : Client
+            MQTT client
+        """
+        names = [w.name for w in self.workflows]
+        print("[%s] Starting in parallel..." % (", ".join(names)))
+        for workflow in self.workflows:
+            workflow.execute(client)
+
+    def dispose(self, client):
+        """
+        Disposes this workflow.
+
+        Parameters
+        ----------
+        client : Client
+            MQTT client
+        """
+        for workflow in self.workflows:
+            workflow.dispose(client)
+
+    def on_message(self, msg):
+        """
+        Processes the message sended by the MQTT server.
+
+        Parameters
+        ----------
+        msg : Message
+            Message from the MQTT topic.
+        """
+        for workflow in self.workflows:
+            workflow.on_message(msg)
+
+    def on_received_status_solved(self, data):
+        # A parallel workflow is solved iff all wrapped workflows are solved
+        pass
+
+    def __on_workflow_failed(self, name, error):
+        self._on_workflow_failed(error)
+
+    def __on_workflow_solved(self, name):
+        self.workflow_solved[name] = True
+        if all(list(self.workflow_solved.values())):
+            self._on_workflow_solved(name)
+
+
+class DoorTargetState(Enum):
+    OPENED = 0
+    CLOSED = 1
+
+
+class DoorWorkflow(Workflow):
+
+    def __init__(self, name, topic, target_state):
+        """
+        Initializes a new instance of this class.
+
+        Parameters
+        ----------
+        name : str
+            Display name of the workflow.
+        topic : str
+            Name of the MQTT topic.
+        target_state: DoorState
+            Target state of the doot (opened/closed)
+        """
+        self.target_state = target_state
+        super().__init__(name, topic)
+
+    def on_received_status_inactive(self, data):
+        """
+        OVERRIDDEN: Door doesn't confirm the state solved.
+        """
+        if data.lower() == self.target_state.name.lower():
+            print("  ==> Door %s" % data.lower())
+            self._on_workflow_solved(self.name)
+        else:
+            super.on_received_status_inactive(data)
