+++ logic/workflow_definition.py
@@ -1,15 +1,17 @@
 from workflow import Workflow
-from parallel_workflow import ParallelWorkflow
+from workflow import ParallelWorkflow
+from workflow import DoorWorkflow
+from workflow import DoorTargetState as ds
 
 
 class WorkflowDefinition:
 
     def get(self):
         return [
-            Workflow("Puzzle 1", "QUEUE_Puzzle_1"),
-            Workflow("Puzzle 2", "QUEUE_Puzzle_2"),
-            Workflow("Puzzle 3", "QUEUE_Puzzle_3"),
-            ParallelWorkflow("Parallel puzzle", "QUEUE_Parallel_Puzzle", [
+            Workflow("Keypad_code_entrance_door", "4/door/entrance/puzzle"),
+            DoorWorkflow("Door_opening", "4/door/entrance", ds.OPENED),
+            Workflow("Globes_server_room", "4/puzzle", {'participants': 4}),
+            ParallelWorkflow("Lab room", [
                 Workflow("Puzzle 4", "QUEUE_Puzzle_4"),
                 Workflow("Puzzle 5", "QUEUE_Puzzle_5")
             ])
